Capitalist Units Attack Base
Also changed xml values of Units

+++ Order_227/Motor2D/Building.h
@@ -50,7 +50,7 @@ public:
 
 
 	
-	int health = 0;
+	float health = 0;
 	int income = 0;
 	Timer incomeTimer;
